// src/types/adguard.d.ts
export interface components {
    schemas: {
        /** @description AdGuard Home server status and configuration */
        ServerStatus: {
            /** @example [
             *       "127.0.0.1"
             *     ] */
            dns_addresses: string[];
            /**
             * Format: uint16
             * @example 53
             */
            dns_port: number;
            /**
             * Format: uint16
             * @example 80
             */
            http_port: number;
            protection_enabled: boolean;
            /** Format: int64 */
            protection_disabled_duration?: number;
            dhcp_available?: boolean;
            running: boolean;
            /** @example v0.123.4 */
            version: string;
            /** @example en */
            language: string;
        };
// ...
        Stats: {
            /**
             * @description Time units
             * @example hours
             * @enum {string}
             */
            time_units?: "hours" | "days";
            /**
             * @description Total number of DNS queries
             * @example 123
             */
            num_dns_queries?: number;
            /**
             * @description Number of requests blocked by filtering rules
             * @example 50
             */
            num_blocked_filtering?: number;
            /**
             * @description Number of requests blocked by safebrowsing module
             * @example 5
             */
            num_replaced_safebrowsing?: number;
            /**
             * @description Number of requests blocked by safesearch module
             * @example 5
             */
            num_replaced_safesearch?: number;
            /**
             * @description Number of blocked adult websites
             * @example 15
             */
            num_replaced_parental?: number;
            /**
             * Format: float
             * @description Average time in seconds on processing a DNS request
             * @example 0.34
             */
            avg_processing_time?: number;
            top_queried_domains?: components["schemas"]["TopArrayEntry"][];
            top_clients?: components["schemas"]["TopArrayEntry"][];
            top_blocked_domains?: components["schemas"]["TopArrayEntry"][];
            /** @description Total number of responses from each upstream. */
            top_upstreams_responses?: components["schemas"]["TopArrayEntry"][];
            /** @description Average processing time in seconds of requests from each upstream.
             *      */
            top_upstreams_avg_time?: components["schemas"]["TopArrayEntry"][];
            dns_queries?: number[];
            blocked_filtering?: number[];
            replaced_safebrowsing?: number[];
            replaced_parental?: number[];
        };
// ...
    };
// ...
}